Add files via upload
In this project we take a dataset in which there are several images of the celebrities like Virat Kohli, Sharapova and Lionel Messi by the help of this data we train the Machine Learnig algorithm so that they can do predict when there new picture is come in the contact. On the basis of this dataset we can detect the picture that this picture is of celebrity, in this we use simple Machine Learning Algorithm rather than the use of the neural network and other Deep Learning model and can detect the face with the accuracy of 80% with SVM model which give best accuracy.